miscellaneous changes to clean up use go modules now use context instead of global channel move types into their own file and add comments etc.
