+++ cmd/pktvisor/go.mod
@@ -1,0 +1,10 @@
+module github.com/ns1/pktvisor/cmd/pktvisor
+
+go 1.15
+
+require (
+	github.com/docopt/docopt-go v0.0.0-20180111231733-ee0de3bc6815
+	github.com/jroimartin/gocui v0.4.0
+	github.com/nsf/termbox-go v0.0.0-20201124104050-ed494de23a00 // indirect
+	github.com/pkg/errors v0.9.1
+)
